Update transifex sync script
This is what runs on the pioneer VM every night, as a cron job.

+++ scripts/update-translations.sh
@@ -1,18 +1,46 @@
-#!/bin/sh
+#!/bin/bash
 
-# XXX this assumes you have push access to github master and transifex, and
-# assume paths. you'll want to hack it before trying to use it yourself
+# Pre-requisites:
+# 1. transifex authentication in ~/.transifexrc
+# 2. github authentication in ~/.ssh/pioneer-transifex.id_rsa
+#    2.1 add to config
+#    Host pioneer-transifex-github
+#      Hostname github.com
+#      User git
+#      IdentityFile ~/.ssh/pioneer-transifex.id_rsa
 
 set -x
 set -e
-export PATH=/usr/local/bin:$PATH
-cd /home/robn/p/pioneer-translation
-git fetch
-git reset --hard origin/master
-tx push -s
-tx pull -a #--minimum-perc=25
-find data/lang -name \*.json -not -name en.json | while read f ; do cat $f | perl -MJSON -e 'undef $/;$j=JSON->new->pretty->utf8->indent->canonical;print $j->encode($j->decode(<>))' | sponge $f ; done
-git add data/lang/*/*.json
-git commit -m 'auto-commit: translation updates'
-git push
+
+cd $HOME/pioneer
+
+if [ "$1" = "init" ]
+then
+    git clone --depth=1 git@pioneer-transifex-github:pioneerspacesim/pioneer
+    cd pioneer
+    git config user.name 'Pioneer Transifex'
+    git config user.email 'pioneer-transifex@pioneerspacesim.net'
+else
+    cd pioneer
+    git fetch
+    git reset --hard origin/master
+
+    # push only our en.json strings to transifex (-s source, -t translation)
+    tx push -s
+
+    # (force) pull translations from transifex to machine
+    # tx pull -a
+    tx pull -t -f
+
+    # remove transifex's indentation:
+    for i in $(find data/lang -name '*.json')
+    do
+        jq -S . < $i > /tmp/${i//\//_}
+        mv /tmp/${i//\//_} $i ;
+    done
+
+    git add data/lang/
+    git commit -m 'auto-commit: translation updates'
+    git push
+fi
 exit 0
